--- client-java/controller/src/main/java/org/evomaster/client/java/controller/internal/db/MongoCommandWithDistance.java
@@ -1,13 +1,0 @@
-package org.evomaster.client.java.controller.internal.db;
-
-public class MongoCommandWithDistance {
-
-    public final Object mongoCommand;
-
-    public final MongoDistanceWithMetrics mongoDistanceWithMetrics;
-
-    public MongoCommandWithDistance(Object mongoCommand, MongoDistanceWithMetrics mongoDistanceWithMetrics) {
-        this.mongoCommand = mongoCommand;
-        this.mongoDistanceWithMetrics = mongoDistanceWithMetrics;
-    }
-}
