+++ client-java/controller/src/test/java/org/evomaster/client/java/controller/internal/db/mongo/MongoHandlerTest.java
@@ -3,8 +3,6 @@ package org.evomaster.client.java.controller.internal.db.mongo;
 import com.mongodb.client.*;
 import org.bson.Document;
 import org.bson.conversions.Bson;
-import org.evomaster.client.java.controller.internal.db.MongoHandler;
-import org.evomaster.client.java.controller.internal.db.MongoCommandWithDistance;
 import org.evomaster.client.java.instrumentation.MongoFindCommand;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
