+++ core/src/test/kotlin/org/evomaster/core/parser/GenePostgresLikeVisitorTest.kt
@@ -76,4 +76,6 @@ class GenePostgresLikeVisitorTest : RegexTestTemplate(){
     fun testPercentWithAlsoEscape(){
         check("\\%%", "%.*")
     }
+
+
 }
